style: cargo clippy --fix --workspace --lib
Auto-fixable clippy lints across all lib targets. No semantic
changes — every fix is one of clippy's safe-by-design suggestions:

  - `std::io::Error::new(Other, _)` → `std::io::Error::other(_)`
  - `length_comparison_to_zero` → `is_empty()`
  - `if let Some(_) = …` collapsed into match
  - `match` to `matches!` macro where appropriate
  - dead branches that always fall through

Same motivation as the preceding `style: cargo fmt --all` commit
(0219c30): the orchestrator-spawned implementer reflexively runs
`cargo clippy --fix` early in its session. Without this commit,
that clippy run would produce a 27-file sprawl that the
implementer then has to spend tokens filtering out before landing
its focused commit. Committing the clippy fixes to main once
makes the implementer's reflex a near-no-op.

Test targets are NOT in this commit because `cargo clippy
--workspace --all-targets` fails on a pre-existing E0063
in pi-ai's test files (`ModelInfo` initializers missing the
`tier` field — same class of pre-existing breakage already
documented in commits 5e08a28 (router_static.rs) and
runner_v1_fixes.rs's prose). Cleaning those up is a separate,
larger effort; for the operational goal (stop the implementer's
fmt/clippy reflex from producing sprawl), the lib-only sweep is
sufficient.

cargo build --workspace clean.

+++ crates/pi-coding-agent/src/native/trajectory/flamegraph.rs
@@ -471,7 +471,7 @@ fn total_tokens(branch: &[SessionEntry]) -> u64 {
     let mut from_usage = 0u64;
     for e in branch {
         if let SessionEntryKind::Usage { usage } = &e.kind {
-            from_usage += usage.input_tokens as u64 + usage.output_tokens as u64;
+            from_usage += usage.input_tokens + usage.output_tokens;
         }
     }
     if from_usage > 0 {
